+++ api/src/Carpool/Service/ResultManager.php
@@ -33,6 +33,7 @@ use App\Carpool\Entity\ResultItem;
 use App\Carpool\Entity\ResultRole;
 use App\Carpool\Repository\AskRepository;
 use App\Carpool\Repository\MatchingRepository;
+use App\Carpool\Repository\ProposalRepository;
 use App\Service\FormatDataManager;
 use App\User\Entity\User;
 use App\User\Repository\ReviewRepository;
@@ -61,6 +62,7 @@ class ResultManager
     private $userReview;
     private $carpoolNoticeableDetourDurationPercent;
     private $carpoolNoticeableDetourDistancePercent;
+    private $proposalRepository;
 
     /**
      * Constructor.
@@ -79,6 +81,7 @@ class ResultManager
         BlockManager $blockManager,
         ReviewRepository $reviewRepository,
         ReviewManager $reviewManager,
+        ProposalRepository $proposalRepository,
         bool $userReview,
         int $carpoolNoticeableDetourDurationPercent,
         int $carpoolNoticeableDetourDistancePercent
@@ -94,6 +97,7 @@ class ResultManager
         $this->userReview = $userReview;
         $this->carpoolNoticeableDetourDurationPercent = $carpoolNoticeableDetourDurationPercent;
         $this->carpoolNoticeableDetourDistancePercent = $carpoolNoticeableDetourDistancePercent;
+        $this->proposalRepository = $proposalRepository;
     }
 
     // set the params
@@ -482,6 +486,7 @@ class ResultManager
                         $blockedRequest = true;
                     }
                 }
+                $matchingProposal = $matching['request']->getProposalRequest();
             }
             if (isset($matching['offer'])) {
                 $user1 = $matching['offer']->getProposalOffer()->getUser();
@@ -493,10 +498,11 @@ class ResultManager
                         $blockedOffer = true;
                     }
                 }
+                $matchingProposal = $matching['offer']->getProposalOffer();
             }
             
             if (!$blockedRequest && !$blockedOffer) {
-                $result = $this->createMatchingResult($proposal, $matchingProposalId, $matching, $return);
+                $result = $this->createMatchingResult($proposal, $matchingProposal, $matching, $return);
                 $results[$matchingProposalId] = $result;
             }
         }
@@ -507,12 +513,12 @@ class ResultManager
      * Create results for a given matching of a proposal
      *
      * @param Proposal $proposal            The proposal
-     * @param integer $matchingProposalId   The proposal that matches
+     * @param Proposal $matchingProposal   The proposal that matches
      * @param array $matching               The array of the matchings of the proposal (an array with the matching proposal as offer and/or request)
      * @param boolean $return               The matching concerns a return (=false if it's the outward)
      * @return Result                       The result object
      */
-    private function createMatchingResult(Proposal $proposal, int $matchingProposalId, array $matching, bool $return)
+    private function createMatchingResult(Proposal $proposal, Proposal $matchingProposal, array $matching, bool $return)
     {
         $result = new Result();
         $result->setId($proposal->getId());
@@ -580,7 +586,7 @@ class ResultManager
             // outward
             $item = new ResultItem();
             // we set the proposalId
-            $item->setProposalId($matchingProposalId);
+            $item->setProposalId($matchingProposal->getId());
             if ($matching['request']->getId() !== Matching::DEFAULT_ID) {
                 $item->setMatchingId($matching['request']->getId());
             }
@@ -1103,7 +1109,7 @@ class ResultManager
             // outward
             $item = new ResultItem();
             // we set the proposalId
-            $item->setProposalId($matchingProposalId);
+            $item->setProposalId($matchingProposal->getId());
             if ($matching['offer']->getId() !== Matching::DEFAULT_ID) {
                 $item->setMatchingId($matching['offer']->getId());
             }
@@ -1613,6 +1619,11 @@ class ResultManager
 
         $result->setCommunities($communities);
 
+        // Check if the matching proposal is owned by the caller (if not anonymous)
+        if (!is_null($proposal->getUser())) {
+            $result->setMyOwn($matchingProposal->getUser()->getId()===$proposal->getUser()->getId());
+        }
+        
         return $result;
     }