+++ api/src/Solidary/Admin/Service/SolidaryVolunteerManager.php
@@ -150,6 +150,7 @@ class SolidaryVolunteerManager
 
             $volunteerStructures[] = [
                 "id" => $solidaryUserStructure->getStructure()->getId(),
+                "userStructureId" => $solidaryUserStructure->getId(),
                 "name" => $solidaryUserStructure->getStructure()->getName(),
                 "status" => $solidaryUserStructure->getStatus()
             ];