+++ client/src/MobicoopBundle/Resources/assets/js/components/base/MHeader.vue
@@ -425,6 +425,10 @@ export default {
     publishButtonAlwaysActive:{
       type: Boolean,
       default:false
+    },
+    gamificationActive:{
+      type: Boolean,
+      default: false
     }
   },
   data () {
@@ -448,9 +452,10 @@ export default {
     } else {
       this.dlocale = this.defaultLocale;
     }
+    this.$store.commit('g/setActive',this.gamificationActive);
   },
   created() {
-    this.$root.token = this.token;
+    this.$store.commit('a/setToken',this.user?.token ? this.user.token : this.token);
     if(this.user){
       localStorage.setItem('X-LOCALE',this.dlocale);
     }