+++ client/src/MobicoopBundle/Resources/assets/js/components/user/profile/Profile.vue
@@ -71,6 +71,7 @@
                   :url-alt-avatar="urlAltAvatar"
                   :avatar-version="avatarVersion"
                   :platform="platform"
+                  :gamification-active="gamificationActive"
                 />
               </v-tab-item>
 
@@ -103,7 +104,17 @@
                   :geo-search-url="geoSearchUrl"
                   :validation-docs-authorized-extensions="validationDocsAuthorizedExtensions"
                 />
-              </v-tab-item>              
+              </v-tab-item>
+              <v-tab
+                v-if="gamificationActive"
+                class="text-subtitle-1"
+              >
+                {{ $t("tabs.myBadges") }}
+              </v-tab>
+              <!-- ACCOUNT -->
+              <v-tab-item v-if="gamificationActive">
+                <Badges />
+              </v-tab-item>                            
             </v-tabs>
           </v-tab-item>
 
@@ -145,7 +156,7 @@ import CarpoolSettings from "@components/user/profile/CarpoolSettings";
 import BankAccount from "@components/user/profile/payment/BankAccount";
 import ProfileSummary from "@components/user/profile/ProfileSummary";
 import ReviewDashboard from "@components/user/profile/review/ReviewDashboard";
-
+import Badges from "@components/user/profile/Badges";
 import {messages_en, messages_fr, messages_eu, messages_nl} from "@translations/components/user/profile/Profile/";
 
 export default {
@@ -165,7 +176,8 @@ export default {
     Carpools,
     BankAccount,
     ProfileSummary,
-    ReviewDashboard
+    ReviewDashboard,
+    Badges
   },
   props: {
     user: {
@@ -223,7 +235,7 @@ export default {
     ageDisplay: {
       type: Boolean,
       default: true
-    },
+    }
   },
   data(){
     return{
@@ -231,6 +243,11 @@ export default {
       publishedAds: {},
       acceptedAds: {}
     }
+  },
+  computed:{
+    gamificationActive(){
+      return this.$store.getters['g/isActive'];
+    },
   },
   mounted(){
     maxios.get(this.$t("getMyCarpools"))