+++ client/src/MobicoopBundle/Resources/assets/js/components/user/profile/UpdateProfile.vue
@@ -675,7 +675,7 @@ export default {
     platform: {
       type: String,
       default: ""
-    }
+    } 
   },
   data() {
     return {